Feature [dev mode] init

+++ init.js
@@ -8,6 +8,10 @@ const yargs = require('yargs')
     help: 'A short package description',
     default: ''
   })
+  .option('dev', {
+    alias: 'i',
+    default: false
+  })
   .argv;
 
 try {
@@ -21,14 +25,18 @@ try {
     files: process.cwd() + '/**',
     from: /{{owner}}/g,
     to: packageOrg
-  })).then(() => {
-    fs.unlinkSync('README.md');
-    fs.renameSync('README.template.md', 'README.md');
+  })).then(() => process.exit(0));
+
+  fs.unlinkSync('README.md');
+  fs.renameSync('README.template.md', 'README.md');
+
+  if (!yargs.dev) {
+    exec('npm un -D replace-in-file yargs').on('exit', (code) => {
+      process.exit(code);
+    });
 
     fs.unlinkSync('init.js');
-
-    process.exit(0);
-  });
+  }
 } catch (e) {
   console.error(`Please, make sure you have provided the package information in the following format:
   node init {github-username-or-org}/{package-name}`);
